feat: Fix Branch CRUD

+++ app/Http/Controllers/BranchController.php
@@ -2,14 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\BranchType;
+use App\Enums\RpsFormat;
+use App\Enums\RpsSimplesNationalOptant;
+use App\Enums\RpsSpecialTaxRegimeInvoice;
+use App\Enums\RpsTaxServiceInvoice;
+use App\Http\Requests\StoreBranchRequest;
+use App\Http\Requests\UpdateBranchRequest;
 use App\Http\Resources\BranchListResource;
 use App\Http\Resources\BranchResource;
 use App\Models\Branch;
 use App\Repositories\BranchRepository;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class BranchController extends Controller
 {
@@ -24,6 +31,7 @@ class BranchController extends Controller
      */
     public function index(): ResourceCollection
     {
+        Gate::authorize('viewAny', Branch::class);
         $branches = $this->branchRepository->paginate(request()->all());
 
         return BranchListResource::collection($branches);
@@ -32,16 +40,9 @@ class BranchController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request): JsonResponse
+    public function store(StoreBranchRequest $request): JsonResponse
     {
-        $data = $request->validate([
-            'name' => 'required|string',
-            'address' => 'required|string',
-            'phone' => 'required|string',
-            'email' => 'required|email',
-        ]);
-        // TODO validate
-
+        $data = $request->validated();
         $branch = $this->branchRepository->create($data);
 
         return response()->json($branch);
@@ -52,27 +53,24 @@ class BranchController extends Controller
      */
     public function show(Branch $branch): JsonResponse
     {
+        Gate::authorize('view', $branch);
+
         return response()->json($branch);
     }
 
     public function edit(Branch $branch): JsonResponse
     {
+        Gate::authorize('update', $branch);
+
         return response()->json(new BranchResource($branch));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Branch $branch): JsonResponse
+    public function update(UpdateBranchRequest $request, Branch $branch): JsonResponse
     {
-        $data = $request->validate([
-            'name' => 'required|string',
-            'address' => 'required|string',
-            'phone' => 'required|string',
-            'email' => 'required|email',
-        ]);
-        // TODO validate
-
+        $data = $request->validated();
         $branch = $this->branchRepository->update($branch, $data);
 
         return response()->json($branch);
@@ -83,6 +81,7 @@ class BranchController extends Controller
      */
     public function destroy(Branch $branch): Response
     {
+        Gate::authorize('delete', $branch);
         $branch->delete();
 
         return response()->noContent();
@@ -90,7 +89,48 @@ class BranchController extends Controller
 
     public function options(): JsonResponse
     {
+        Gate::authorize('viewAny', Branch::class);
         $options = $this->branchRepository->options();
+
+        return response()->json($options);
+    }
+
+    public function typeOptions(): JsonResponse
+    {
+        Gate::authorize('viewAny', Branch::class);
+        $options = BranchType::options();
+
+        return response()->json($options);
+    }
+
+    public function rpsFormatOptions(): JsonResponse
+    {
+        Gate::authorize('viewAny', Branch::class);
+        $options = RpsFormat::options();
+
+        return response()->json($options);
+    }
+
+    public function rpsTaxServiceInvoiceOptions(): JsonResponse
+    {
+        Gate::authorize('viewAny', Branch::class);
+        $options = RpsTaxServiceInvoice::options();
+
+        return response()->json($options);
+    }
+
+    public function rpsSpecialTaxRegimeInvoiceOptions(): JsonResponse
+    {
+        Gate::authorize('viewAny', Branch::class);
+        $options = RpsSpecialTaxRegimeInvoice::options();
+
+        return response()->json($options);
+    }
+
+    public function rpsSimplesNationalOptantOptions(): JsonResponse
+    {
+        Gate::authorize('viewAny', Branch::class);
+        $options = RpsSimplesNationalOptant::options();
 
         return response()->json($options);
     }
